Split up annotate_models spec file
