+++ spec/lib/annotate/annotate_models_spec_get_patterns.rb
@@ -1,0 +1,111 @@
+# encoding: utf-8
+require 'annotate/annotate_models'
+require 'annotate/active_record_patch'
+require 'active_support/core_ext/string'
+require 'files'
+require 'tmpdir'
+
+RSpec.describe AnnotateModels do
+  MAGIC_COMMENTS = [
+    '# encoding: UTF-8',
+    '# coding: UTF-8',
+    '# -*- coding: UTF-8 -*-',
+    '#encoding: utf-8',
+    '# encoding: utf-8',
+    '# -*- encoding : utf-8 -*-',
+    "# encoding: utf-8\n# frozen_string_literal: true",
+    "# frozen_string_literal: true\n# encoding: utf-8",
+    '# frozen_string_literal: true',
+    '#frozen_string_literal: false',
+    '# -*- frozen_string_literal : true -*-'
+  ].freeze unless const_defined?(:MAGIC_COMMENTS)
+
+  def mock_index(name, params = {})
+    double('IndexKeyDefinition',
+           name:          name,
+           columns:       params[:columns] || [],
+           unique:        params[:unique] || false,
+           orders:        params[:orders] || {},
+           where:         params[:where],
+           using:         params[:using])
+  end
+
+  def mock_foreign_key(name, from_column, to_table, to_column = 'id', constraints = {})
+    double('ForeignKeyDefinition',
+           name:         name,
+           column:       from_column,
+           to_table:     to_table,
+           primary_key:  to_column,
+           on_delete:    constraints[:on_delete],
+           on_update:    constraints[:on_update])
+  end
+
+  def mock_connection(indexes = [], foreign_keys = [])
+    double('Conn',
+           indexes:      indexes,
+           foreign_keys: foreign_keys,
+           supports_foreign_keys?: true)
+  end
+
+  def mock_class(table_name, primary_key, columns, indexes = [], foreign_keys = [])
+    options = {
+      connection:       mock_connection(indexes, foreign_keys),
+      table_exists?:    true,
+      table_name:       table_name,
+      primary_key:      primary_key,
+      column_names:     columns.map { |col| col.name.to_s },
+      columns:          columns,
+      column_defaults:  Hash[columns.map { |col| [col.name, col.default] }],
+      table_name_prefix: ''
+    }
+
+    double('An ActiveRecord class', options)
+  end
+
+  def mock_column(name, type, options = {})
+    default_options = {
+      limit: nil,
+      null: false,
+      default: nil,
+      sql_type: type
+    }
+
+    stubs = default_options.dup
+    stubs.merge!(options)
+    stubs[:name] = name
+    stubs[:type] = type
+
+    double('Column', stubs)
+  end
+
+  describe '.get_patterns' do
+    subject { AnnotateModels.get_patterns(options, pattern_type) }
+
+    context 'when pattern_type is "additional_file_patterns"' do
+      let(:pattern_type) { 'additional_file_patterns' }
+
+      context 'when additional_file_patterns is specified in the options' do
+        let(:additional_file_patterns) do
+          [
+            '/%PLURALIZED_MODEL_NAME%/**/*.rb',
+            '/bar/%PLURALIZED_MODEL_NAME%/*_form'
+          ]
+        end
+
+        let(:options) { { additional_file_patterns: additional_file_patterns } }
+
+        it 'returns additional_file_patterns in the argument "options"' do
+          is_expected.to eq(additional_file_patterns)
+        end
+      end
+
+      context 'when additional_file_patterns is not specified in the options' do
+        let(:options) { {} }
+
+        it 'returns an empty array' do
+          is_expected.to eq([])
+        end
+      end
+    end
+  end
+end
