+++ spec/lib/annotate/annotate_models_spec_parse_options.rb
@@ -1,0 +1,148 @@
+# encoding: utf-8
+require 'annotate/annotate_models'
+require 'annotate/active_record_patch'
+require 'active_support/core_ext/string'
+require 'files'
+require 'tmpdir'
+
+RSpec.describe AnnotateModels do
+  MAGIC_COMMENTS = [
+    '# encoding: UTF-8',
+    '# coding: UTF-8',
+    '# -*- coding: UTF-8 -*-',
+    '#encoding: utf-8',
+    '# encoding: utf-8',
+    '# -*- encoding : utf-8 -*-',
+    "# encoding: utf-8\n# frozen_string_literal: true",
+    "# frozen_string_literal: true\n# encoding: utf-8",
+    '# frozen_string_literal: true',
+    '#frozen_string_literal: false',
+    '# -*- frozen_string_literal : true -*-'
+  ].freeze unless const_defined?(:MAGIC_COMMENTS)
+
+  def mock_index(name, params = {})
+    double('IndexKeyDefinition',
+           name:          name,
+           columns:       params[:columns] || [],
+           unique:        params[:unique] || false,
+           orders:        params[:orders] || {},
+           where:         params[:where],
+           using:         params[:using])
+  end
+
+  def mock_foreign_key(name, from_column, to_table, to_column = 'id', constraints = {})
+    double('ForeignKeyDefinition',
+           name:         name,
+           column:       from_column,
+           to_table:     to_table,
+           primary_key:  to_column,
+           on_delete:    constraints[:on_delete],
+           on_update:    constraints[:on_update])
+  end
+
+  def mock_connection(indexes = [], foreign_keys = [])
+    double('Conn',
+           indexes:      indexes,
+           foreign_keys: foreign_keys,
+           supports_foreign_keys?: true)
+  end
+
+  def mock_class(table_name, primary_key, columns, indexes = [], foreign_keys = [])
+    options = {
+      connection:       mock_connection(indexes, foreign_keys),
+      table_exists?:    true,
+      table_name:       table_name,
+      primary_key:      primary_key,
+      column_names:     columns.map { |col| col.name.to_s },
+      columns:          columns,
+      column_defaults:  Hash[columns.map { |col| [col.name, col.default] }],
+      table_name_prefix: ''
+    }
+
+    double('An ActiveRecord class', options)
+  end
+
+  def mock_column(name, type, options = {})
+    default_options = {
+      limit: nil,
+      null: false,
+      default: nil,
+      sql_type: type
+    }
+
+    stubs = default_options.dup
+    stubs.merge!(options)
+    stubs[:name] = name
+    stubs[:type] = type
+
+    double('Column', stubs)
+  end
+
+  describe '.parse_options' do
+    let(:options) do
+      {
+        root_dir: '/root',
+        model_dir: 'app/models,app/one,  app/two   ,,app/three',
+        skip_subdirectory_model_load: false
+      }
+    end
+
+    before :each do
+      AnnotateModels.send(:parse_options, options)
+    end
+
+    describe '@root_dir' do
+      subject do
+        AnnotateModels.instance_variable_get(:@root_dir)
+      end
+
+      it 'sets @root_dir' do
+        is_expected.to eq('/root')
+      end
+    end
+
+    describe '@model_dir' do
+      subject do
+        AnnotateModels.instance_variable_get(:@model_dir)
+      end
+
+      it 'separates option "model_dir" with commas and sets @model_dir as an array of string' do
+        is_expected.to eq(['app/models', 'app/one', 'app/two', 'app/three'])
+      end
+    end
+
+    describe '@skip_subdirectory_model_load' do
+      subject do
+        AnnotateModels.instance_variable_get(:@skip_subdirectory_model_load)
+      end
+
+      context 'option is set to true' do
+        let(:options) do
+          {
+            root_dir: '/root',
+            model_dir: 'app/models,app/one,  app/two   ,,app/three',
+            skip_subdirectory_model_load: true
+          }
+        end
+
+        it 'sets skip_subdirectory_model_load to true' do
+          is_expected.to eq(true)
+        end
+      end
+
+      context 'option is set to false' do
+        let(:options) do
+          {
+            root_dir: '/root',
+            model_dir: 'app/models,app/one,  app/two   ,,app/three',
+            skip_subdirectory_model_load: false
+          }
+        end
+
+        it 'sets skip_subdirectory_model_load to false' do
+          is_expected.to eq(false)
+        end
+      end
+    end
+  end
+end
